Thinned out clients, slight refactor
