+++ metis-preview/metis-preview-client/pom.xml
@@ -10,10 +10,24 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>metis-preview-client</artifactId>
+    <packaging>pom</packaging>
+    <modules>
+        <module>../metis-preview-common</module>
+    </modules>
     <dependencies>
         <dependency>
             <groupId>eu.europeana.metis</groupId>
-            <artifactId>metis-preview-service</artifactId>
+            <artifactId>metis-preview-common</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-web</artifactId>
+            <version>4.3.1.RELEASE</version>
+        </dependency>
+        <dependency>
+            <groupId>eu.europeana.metis</groupId>
+            <artifactId>metis-common</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
