+++ metis-preview/metis-preview-common/pom.xml
@@ -3,21 +3,20 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>metis-framework</artifactId>
+        <artifactId>metis-preview-client</artifactId>
         <groupId>eu.europeana.metis</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>metis-preview</artifactId>
-    <packaging>pom</packaging>
-    <modules>
-        <module>metis-preview-common</module>
-        <module>metis-preview-service</module>
-        <module>metis-preview-rest</module>
-        <module>metis-preview-client</module>
-        <module>metis-preview-persistence</module>
-    </modules>
+    <artifactId>metis-preview-common</artifactId>
+    <dependencies>
+        <dependency>
+            <groupId>eu.europeana.metis</groupId>
+            <artifactId>validation-common</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+    </dependencies>
 
 
 </project>
