+++ metis-preview/metis-preview-service/src/main/java/eu/europeana/metis/preview/service/executor/ValidationTask.java
@@ -6,8 +6,8 @@ import eu.europeana.corelib.edm.utils.MongoConstructor;
 import eu.europeana.corelib.solr.bean.impl.FullBeanImpl;
 import eu.europeana.metis.dereference.service.xslt.XsltTransformer;
 import eu.europeana.metis.identifier.RestClient;
+import eu.europeana.metis.preview.model.ExtendedValidationResult;
 import eu.europeana.metis.preview.persistence.RecordDao;
-import eu.europeana.metis.preview.service.ExtendedValidationResult;
 import eu.europeana.validation.client.ValidationClient;
 import eu.europeana.validation.model.ValidationResult;
 import org.apache.commons.io.FileUtils;
