+++ metis-validation/validation-client/pom.xml
@@ -43,11 +43,6 @@
             <version>${version.spring}</version>
         </dependency>
         <dependency>
-            <groupId>org.jetbrains</groupId>
-            <artifactId>annotations</artifactId>
-            <version>RELEASE</version>
-        </dependency>
-        <dependency>
             <groupId>eu.europeana.metis</groupId>
             <artifactId>metis-common</artifactId>
             <version>1.0-SNAPSHOT</version>
